Extract code specific to 'reprojucer' mode to "reprojucer.hpp"
The diff of this commit is easier to read when using a tool that shows
similarity between files, such as gitk, because "reprojucer.hpp" is
almost a copy of "main.cpp" (similarity index 92%).

The copyright notices for Matthieu Talbot, Florian Goltz, and Johannes
Elliesen were moved from "main.cpp" to "reprojucer.hpp" since their
respective contributions also moved.
